refactor: api 함수 모듈화 하기
