+++ src/types/mock.ts
@@ -1,16 +1,4 @@
-import axios from 'axios';
-
 import type { GoodsData, ThemeData } from '.';
-
-export const fetchThemeData = async (): Promise<ThemeData[]> => {
-  try {
-    const response = await axios.get('https://kakao-tech-campus-mock-server.vercel.app/api/v1/themes');
-    return response.data.themes;
-  } catch (error) {
-    console.error('Failed to fetch theme data:', error);
-    return [];
-  }
-};
 
 export const ThemeMockList: ThemeData[] = [];
